refactoring

+++ controllers/card.js
@@ -1,7 +1,8 @@
-const Card = require('../models/Card');
-const BadRequestError = require('../errors/BadRequestError');
-const NotFoundError = require('../errors/NotFoundError');
-const ServerError = require('../errors/ServerError');
+const Card = require("../models/Card");
+const BadRequestError = require("../errors/BadRequestError");
+const NotFoundError = require("../errors/NotFoundError");
+const ForbiddenError = require("../errors/ForbiddenError");
+const ServerError = require("../errors/ServerError");
 
 module.exports.getAllCards = (req, res, next) => {
   Card.find({})
@@ -21,8 +22,10 @@ module.exports.createCard = (req, res, next) => {
   Card.create({ name, link, owner: ownerId })
     .then((dataCard) => res.status(201).send({ data: dataCard._id }))
     .catch((err) => {
-      if (err.name === 'ValidationError') {
-        throw new BadRequestError('Переданы некорректные данные при создании карточки');
+      if (err.name === "ValidationError") {
+        throw new BadRequestError(
+          "Переданы некорректные данные при создании карточки"
+        );
       }
       throw new ServerError();
     })
@@ -30,16 +33,30 @@ module.exports.createCard = (req, res, next) => {
 };
 
 module.exports.deleteByIdCard = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.id)
+  Card.findById(req.params.id)
     .orFail(() => {
-      throw new Error('NotFound');
+      throw new Error("NotFound");
     })
-    .then((dataCard) => res.status(200).send(dataCard))
+    .then((dataCard) => {
+      if (dataCard.owner.toString() !== req.user._id) {
+        throw new ForbiddenError("Нет прав");
+      }
+      Card.findByIdAndRemove(req.params.id)
+        .orFail(() => {
+          throw new Error("NotFound");
+        })
+        .then((dataCard) => {
+          if (dataCard.owner.toString() !== req.user._id) {
+            throw new ForbiddenError("Нет прав");
+          }
+          res.status(200).send(dataCard);
+        });
+    })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        throw new BadRequestError('Карточка с указанным _id не найдена');
-      } else if (err.message === 'NotFound') {
-        throw new NotFoundError('Карточка с указанным _id не найдена');
+      if (err.name === "CastError") {
+        throw new BadRequestError("Карточка с указанным _id не найдена");
+      } else if (err.message === "NotFound") {
+        throw new NotFoundError("Карточка с указанным _id не найдена");
       }
       throw new ServerError();
     })
@@ -50,17 +67,17 @@ module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
-    { new: true },
+    { new: true }
   )
     .orFail(() => {
-      throw new Error('NotFound');
+      throw new Error("NotFound");
     })
     .then((dataCard) => res.status(200).send({ data: dataCard }))
     .catch((err) => {
-      if (err.name === 'CastError') {
-        throw new BadRequestError('Передан несуществующий _id карточки');
-      } else if (err.message === 'NotFound') {
-        throw new NotFoundError('Карточка с указанным _id не найдена');
+      if (err.name === "CastError") {
+        throw new BadRequestError("Передан несуществующий _id карточки");
+      } else if (err.message === "NotFound") {
+        throw new NotFoundError("Карточка с указанным _id не найдена");
       }
       throw new ServerError();
     })
@@ -71,17 +88,17 @@ module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
-    { new: true },
+    { new: true }
   )
     .orFail(() => {
-      throw new Error('NotFound');
+      throw new Error("NotFound");
     })
     .then((dataCard) => res.status(200).send({ data: dataCard }))
     .catch((err) => {
-      if (err.name === 'CastError') {
-        throw new BadRequestError('Передан несуществующий _id карточки');
-      } else if (err.message === 'NotFound') {
-        throw new NotFoundError('Карточка с указанным _id не найдена');
+      if (err.name === "CastError") {
+        throw new BadRequestError("Передан несуществующий _id карточки");
+      } else if (err.message === "NotFound") {
+        throw new NotFoundError("Карточка с указанным _id не найдена");
       }
       throw new ServerError();
     })
